refactor: рефакторинг страницы Questions по замечаниям
